requirements.txt
litellm
python-telegram-bot[webhooks]
gitpython
mcp
python-dotenv
mem0ai
duckduckgo-search
apscheduler
